pyproject.toml
[project]
name = "contextly"
version = "0.0.1"

authors = [
    { name = "Ivan Baklanov", email = "ivanbaklan@gmail.com" },
]

requires-python = "==3.11.*"

dependencies = [
    "accelerate>=1.2.0",
    "aerich>=0.8.0",
    "certifi>=2024.8.30",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "huggingface-hub>=0.26.5",
    "jinja2>=3.1.4",
    "loguru>=0.7.3",
    "openai-whisper>=20240930",
    "pip-system-certs>=4.0",
    "protobuf>=5.29.1",
    "pytest-asyncio>=0.25.0",
    "python-multipart>=0.0.19",
    "sentencepiece>=0.2.0",
    "tortoise-orm>=0.21.7",
    "asyncpg>=0.29.0",
    "transformers>=4.47.0",
    "uvicorn>=0.32.1",
    "yt-dlp>=2024.12.6",
    "passlib>=1.7.4",
    "bcrypt>=4.2.1",
    "ffmpeg-python>=0.2.0",
    "python-jose>=3.3.0",
]


[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "pytest==8.3.*",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.25.0",
    "asgi-lifespan>=2.1.0",
]
